fix build, move chat mocks to kbtest
